requirements.txt
fastapi==0.109.2
python-multipart==0.0.9
google-generativeai==0.3.2
python-dotenv==1.0.1
uvicorn==0.27.1
python-docx==1.0.1
pdfplumber==0.10.3